move log_queue to mp_queue
